Fix bug regarding justPressedEquals
The issue was that justPressedEquals being made True after pressing
equals did allow a new number to be typed in and populate the display,
but meant that every single time this did happen, reset() was triggered.
This is fine as long as, the next time you press an operate after equals,
justPressedEquals is made false again, so the next time you type in an
operand reset is not called and you can continue operations. So:

- Make clicking an operator button set justPressedEquals to false