Fix bug preventing choice of operator being changeable
To do this, maybe make it so that, when an operator button has just been
pressed, "justPressedOperator" is made true. Next, when you press an operator and
this is true, all that should happen is that the click should replace
queuedOperator with the new operator, and do this before the check for whether
the secondOperand is null. When you type in a new digit or equals or trigger
reset(), however, justPressedOperator should be made false if it is true.